% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_spec.R
\name{col_spec_compare}
\alias{col_spec_compare}
\title{Determine amount of disagreement}
\usage{
col_spec_compare(x, y)
}
\arguments{
\item{x}{\code{\link[readr:cols]{readr::cols()}} column-specification, or \code{data.frame}}

\item{y}{\code{\link[readr:cols]{readr::cols()}} column-specification, or \code{data.frame}}
}
\value{
object with S3 class \code{col_spec_diff}
}
\description{
Determine amount of disagreement
}
