% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cols_compare.R
\name{cols_compare}
\alias{cols_compare}
\title{Determine amount of disagreement}
\usage{
cols_compare(.x, .y)
}
\arguments{
\item{.x}{\code{\link[readr:cols]{readr::cols()}} column-specification, or \code{data.frame}}

\item{.y}{\code{\link[readr:cols]{readr::cols()}} column-specification, or \code{data.frame}}
}
\value{
object with S3 class \code{cols_diff}
}
\description{
Determine amount of disagreement
}
