% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{diamonds_meta}
\alias{diamonds_meta}
\title{Diamonds metadata}
\format{object with S3 class \code{stw_meta}}
\source{
\url{https://CRAN.R-project.org/package=ggplot2}
}
\usage{
diamonds_meta
}
\description{
The diamonds dataset comes from ggplot2. This is its metadata.
}
\keyword{datasets}
