% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{read_stw}
\alias{read_stw}
\title{Read Steward Data Frame, Metadata together from File}
\usage{
read_stw(file)
}
\arguments{
\item{file}{}
}
\description{
Read Steward Data Frame, Metadata together from File
}
