% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stw_dataset.R
\name{stw_dataset}
\alias{stw_dataset}
\title{Create new dataset object}
\usage{
stw_dataset(data, meta = NULL, ...)
}
\arguments{
\item{data}{\code{data.frame} to be documented}

\item{meta}{Object with S3 class \code{stw_meta}}

\item{...}{Name-value pairs, passed on to \code{\link[=stw_mutate_meta]{stw_mutate_meta()}}}
}
\value{
object with S3 class \code{stw_dataset}
}
\description{
Create new dataset object
}
