% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{stw_df_skeleton}
\alias{stw_df_skeleton}
\title{Initiate a Steward Data Frame with Empty Metadata}
\usage{
stw_df_skeleton(.data)
}
\arguments{
\item{.data}{}
}
\value{
\code{stw_df}
}
\description{
Initiate a Steward Data Frame with Empty Metadata
}
