% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stw_dict.R
\name{stw_dict}
\alias{stw_dict}
\title{Create a new data-dictionary object}
\usage{
stw_dict(dictionary)
}
\arguments{
\item{dictionary}{\code{data.frame} that has columns \code{name}, \code{type}, \code{description}}
}
\value{
object with S3 class \code{stw_meta}
}
\description{
Create a new data-dictionary object
}
