% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{stw_impose}
\alias{stw_impose}
\title{Impose Metadata Objects on a Data Frame}
\usage{
stw_impose(.data, .stw_meta, .stw_var_meta)
}
\arguments{
\item{.stw_var_meta}{}
}
\description{
Impose Metadata Objects on a Data Frame
}
