% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stw_meta.R
\name{stw_meta_env}
\alias{stw_meta_env}
\title{Create stw_meta from list-like object}
\usage{
stw_meta_env(env)
}
\arguments{
\item{env}{a list-like object with elements \code{name}, \code{title}, etc.}
}
\value{
object with S3 class \code{stw_meta}
}
\description{
Create stw_meta from list-like object
}
