% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{stw_meta_skeleton}
\alias{stw_meta_skeleton}
\title{Make a Metadata Skeleton from a Data Frame}
\usage{
stw_meta_skeleton(.data, title = deparse(substitute(.data)))
}
\arguments{
\item{name}{}
}
\description{
Make a Metadata Skeleton from a Data Frame
}
