% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stw_mutate_meta.R
\name{stw_mutate_meta}
\alias{stw_mutate_meta}
\alias{stw_mutate_meta.default}
\alias{stw_mutate_meta.stw_meta}
\alias{stw_mutate_meta.stw_dataset}
\title{Modify metadata-object}
\usage{
stw_mutate_meta(...)

\method{stw_mutate_meta}{default}(...)

\method{stw_mutate_meta}{stw_meta}(meta, ...)

\method{stw_mutate_meta}{stw_dataset}(dataset, ...)
}
\arguments{
\item{...}{Name-value pairs, passed on to \code{\link[=stw_mutate_meta]{stw_mutate_meta()}}}

\item{meta}{Object with S3 class \code{stw_meta}, contains meta-data}

\item{dataset}{Object with S3 class \code{stw_dataset}, a data frame with attached meta-data}
}
\value{
modified copy of \code{meta} or \code{dataset}
}
\description{
Use this function to modify a metadata object,
using syntax along the same lines as \code{\link[dplyr:mutate]{dplyr::mutate()}}.
}
\details{
You can send a metadata (\code{stw_meta}) object or a dataset (\code{stw_dataset})
object, and get back an object of the same class. In other words, if
you call this function using a \code{stw_dataset} object, it will modify the
metadata, then return a modified copy of the \code{stw_dataset} object.
}
\examples{
stw_mutate_meta(diamonds_meta, title = "New title")

}
