% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stw_read_csv.R
\name{stw_read_csv}
\alias{stw_read_csv}
\title{Create metadata-object from CSV}
\usage{
stw_read_csv(file)
}
\arguments{
\item{file}{location of csv file}
}
\value{
object with S3 class \code{stw_meta}
}
\description{
Create metadata-object from CSV
}
\examples{
stw_read_csv(system.file("metadata/diamonds.csv", package = "steward"))
}
