% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stw_read_yaml.R
\name{stw_read_yaml}
\alias{stw_read_yaml}
\title{Create metadata object from YAML}
\usage{
stw_read_yaml(file)
}
\arguments{
\item{file}{location of yaml file}
}
\value{
object with S3 class \code{stw_meta}
}
\description{
Create metadata object from YAML
}
\examples{
stw_read_yaml(system.file("metadata/diamonds.yaml", package = "steward"))


}
