% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stw_to_table.R
\name{stw_to_table}
\alias{stw_to_table}
\alias{stw_to_table.default}
\alias{stw_to_table.stw_meta}
\title{Output to gt (table) format}
\usage{
stw_to_table(meta, ...)

\method{stw_to_table}{default}(meta, ...)

\method{stw_to_table}{stw_meta}(meta, ...)
}
\arguments{
\item{meta}{object with S3 class \code{stw_meta}}

\item{...}{additional arguments (not used)}
}
\value{
A gt table
}
\description{
Output to gt (table) format
}
\examples{
stw_to_table(diamonds_meta)

}
