% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stw_to_yaml.R
\name{stw_to_yaml}
\alias{stw_to_yaml}
\alias{stw_to_yaml.default}
\alias{stw_to_yaml.stw_meta}
\alias{stw_use_yaml}
\alias{stw_write_yaml}
\title{Output to YAML format}
\usage{
stw_to_yaml(meta, ...)

\method{stw_to_yaml}{default}(meta, ...)

\method{stw_to_yaml}{stw_meta}(meta, ...)

stw_use_yaml(meta)

stw_write_yaml(meta, file)
}
\arguments{
\item{meta}{object with S3 class \code{stw_meta}}

\item{...}{additional arguments (not used)}

\item{file}{\code{character} path of file to write}
}
\value{
\describe{
\item{\code{stw_to_yaml()}}{\code{character}, YAML string}
\item{\code{stw_use_yaml()}}{\code{invisible(meta)}, called for side-effects}
\item{\code{stw_write_yaml()}}{\code{invisible(meta)}, called for side-effects}
}
}
\description{
Use these functions to convert your metadata to YAML format. The function
\code{stw_to_yaml()} returns a YAML string, \code{stw_use_yaml()} copies a YAML string
to your clipboard, and \code{str_write_yaml()} writes a YAML string to a file.
}
\examples{
cat(stw_to_yaml(diamonds_meta))
stw_write_yaml(diamonds_meta, tempfile(fileext = ".yml"))
\dontrun{
# not run because it copies text to the clipboard
stw_use_yaml(diamonds_meta)
}

}
