% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stw_use_data.R
\name{stw_use_data}
\alias{stw_use_data}
\title{Create and document package-data}
\usage{
stw_use_data(
  ...,
  file_doc = NULL,
  keep_steward = FALSE,
  overwrite = FALSE,
  compress = "bzip2",
  version = 2
)
}
\arguments{
\item{...}{Unquoted names of existing objects to save.}

\item{file_doc}{\code{character} name of the Roxygen-documentation file to be
written to the package's \code{R} directory; default is
\code{"data-{name-of-first-dataset}.R"}}

\item{keep_steward}{\code{logical} indicates to keep the steward classes and
attributes when writing the dataset to the package.}

\item{overwrite}{By default, \code{use_data()} will not overwrite existing
files. If you really want to do so, set this to \code{TRUE}.}

\item{compress}{Choose the type of compression used by \code{\link[=save]{save()}}.
Should be one of "gzip", "bzip2", or "xz".}

\item{version}{The serialization format version to use. The default, 2, was
the default format from R 1.4.0 to 3.5.3. Version 3 became the default from
R 3.6.0 and can only be read by R versions 3.5.0 and higher.}
}
\value{
\code{invisible(NULL)}, called for side-effects
}
\description{
This function wraps the functionality of \code{\link[usethis:use_data]{usethis::use_data()}} and
\code{\link[=stw_write_roxygen]{stw_write_roxygen()}} to write a data-frame (or tibble) to your package data,
as well as to write the Roxygen-documentation to a file in your package's \code{R}
directory.
}
