% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{stw_var_meta_skeleton}
\alias{stw_var_meta_skeleton}
\title{Make a Variable Metadata Skeleton from a Data Frame}
\usage{
stw_var_meta_skeleton(.data)
}
\arguments{
\item{.data}{}
}
\description{
Make a Variable Metadata Skeleton from a Data Frame
}
