% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbosity.R
\name{with_verbosity}
\alias{with_verbosity}
\alias{verbosely}
\title{Run code with verbosity}
\usage{
with_verbosity(verbosity = c("error", "info", "all", "none"), code)

verbosely(.f, verbosity = c("error", "info", "all", "none"))
}
\arguments{
\item{verbosity}{\code{character} value to set for verbosity option}

\item{code}{Code to be run using the verbosity option}

\item{.f}{\code{function} to wrap using \code{verbosity}}
}
\value{
\describe{
\item{with_verbosity}{The results of the evaluation of the \code{code} argument.}
\item{verbosely}{\code{function} with same arguments as \code{.f}, which will run
using the \code{verbosity} argument}
}
}
\description{
Use these functions to wrap code in a \code{verbosity} option. You can
wrap blocks of code using \code{with_verbosity()}; you can use the adverb
\code{verbosely()} to modify another function so it acts with \code{verbosity}.
}
\details{
Keep in mind that only the "closest" invocation to the code will be
effective; the allowable values for \code{verbosity} are:

\describe{
\item{\code{"error"}}{reports results of all checks that failed}
\item{\code{"info"}}{reports results of all checks that failed,
and that find missing optional information}
\item{\code{"all"}}{reports results of all checks}
\item{\code{"none"}}{reports no results}
}

TODO: talk about implementation, \code{steward.verbosity}
}
\examples{
with_verbosity("all", stw_meta(diamonds_meta))
verbosely(stw_meta, "all")(diamonds_meta)
diamonds_meta \%>\% verbosely(stw_meta, "all")()

}
