% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{write_stw}
\alias{write_stw}
\title{Write Steward Data Frame or Metadata to File}
\usage{
write_stw(.stw_df, file, which = c("data", "meta", "var_meta"), ...)
}
\arguments{
\item{...}{}
}
\description{
Write Steward Data Frame or Metadata to File
}
