% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stw_validate.R
\name{stw_check}
\alias{stw_check}
\alias{stw_check.default}
\alias{stw_check.stw_dict}
\alias{stw_validate}
\alias{stw_validate.default}
\title{Check a steward object}
\usage{
stw_check(...)

\method{stw_check}{default}(...)

\method{stw_check}{stw_dict}(dict, verbosity = c("none", "error", "info",
  "all"), ...)

stw_validate(...)

\method{stw_validate}{default}(...)
}
\arguments{
\item{...}{other arguments (not used)}

\item{dict}{object with S3 class \code{stw_dict},
contains dataset data-dictionary; see \code{\link[=stw_dict]{stw_dict()}}.}

\item{verbosity}{\code{character}, determines the amount of feedback, see Details.}
}
\value{
modified copy of \code{dict} or \code{meta}
}
\description{
Use this function to determine the validity of a \code{stw_dict} or \code{stw_meta} object.
}
\details{
You can specify the `verbosity`` of the check:

\describe{
\item{\code{"all"}}{reports results of all checks}
\item{\code{"info"}}{reports results of all checks that failed,
and that have missing optional information}
\item{\code{"error"}}{reports results of all checks that failed}
\item{\code{"none"}}{reports no results}
}

The \code{"none"} option may not seem intuitive, it is for internal use.
These functions set an internal attribute of the object to inicate its
validity.

For a \code{stw_dict} object each \code{name} nust be unique and non-trivial,
each \code{description} must be non-trivial, \code{type} is optional.

A \code{stw_meta} object must have a non-trivial \code{name}, and \code{dict}.
It may have a \code{title}, \code{description}, \code{source}, \code{n_row}, and \code{n_col}.
}
\examples{
stw_check(diamonds_meta$dict)

}
