% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stw_validate.R
\name{stw_check}
\alias{stw_check}
\alias{stw_check.default}
\alias{stw_check.stw_dict}
\alias{stw_check.stw_meta}
\alias{stw_validate}
\alias{stw_validate.default}
\alias{stw_validate.stw_dict}
\alias{stw_validate.stw_meta}
\alias{stw_validate.stw_dataset}
\title{Check steward object}
\usage{
stw_check(...)

\method{stw_check}{default}(...)

\method{stw_check}{stw_dict}(dict, verbosity = c("error", "info", "all",
  "none"), ...)

\method{stw_check}{stw_meta}(meta, verbosity = c("error", "info", "all",
  "none"), ...)

stw_validate(...)

\method{stw_validate}{default}(...)

\method{stw_validate}{stw_dict}(dict, verbosity = c("error", "info",
  "all"), ...)

\method{stw_validate}{stw_meta}(meta, verbosity = c("error", "info",
  "all"), ...)

\method{stw_validate}{stw_dataset}(dataset, verbosity = c("error",
  "info", "all"), ...)
}
\arguments{
\item{...}{other arguments (not used)}

\item{dict}{Object with S3 class \code{stw_dict}, contains data-dictionary}

\item{verbosity}{\code{character}, determines the amount of feedback, see Details.}

\item{meta}{Object with S3 class \code{stw_meta}, contains meta-data}

\item{dataset}{Object with S3 class \code{stw_dataset}, a data frame with attached meta-data}
}
\value{
modified copy of \code{dict} or \code{meta}
}
\description{
Use these function to determine the validity of a \code{stw_dict} or \code{stw_meta} object;
\code{stw_check()} tells you about the object, \code{stw_validate()} additionally throws
an error if warranted. To support piping, these functions return their first
argument invisibly.
}
\details{
You can specify the \code{verbosity}:

\describe{
\item{\code{"error"}}{reports results of all checks that failed}
\item{\code{"info"}}{reports results of all checks that failed,
and that find missing optional information}
\item{\code{"all"}}{reports results of all checks}
\item{\code{"none"}}{reports no results}
}

The \code{"none"} option may not seem intuitive, it is for internal use.
These functions set an internal attribute of the object to inicate its
validity.

For a \code{stw_dict} object, each \code{name} nust be unique and non-trivial,
each \code{description} must be non-trivial, \code{type} is optional.

For a \code{stw_meta} object, it must have a non-trivial \code{name}, and \code{dict}.
It may have a \code{title}, \code{description}, \code{source}, \code{n_row}, and \code{n_col}.
}
\examples{
stw_check(diamonds_meta$dict)

}
