% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stw_dataset.R
\name{stw_dataset}
\alias{stw_dataset}
\alias{stw_dataset.default}
\alias{stw_dataset.data.frame}
\alias{stw_dataset.stw_dataset}
\title{Create new dataset object}
\usage{
stw_dataset(...)

\method{stw_dataset}{default}(...)

\method{stw_dataset}{data.frame}(data, meta = NULL, ...)

\method{stw_dataset}{stw_dataset}(dataset, meta = NULL, ...)
}
\arguments{
\item{...}{Name-value pairs, passed on to \code{\link[=stw_mutate_meta]{stw_mutate_meta()}}}

\item{data}{\code{data.frame} to be documented}

\item{meta}{Object with S3 class \code{stw_meta}, contains meta-data}

\item{dataset}{Object with S3 class \code{stw_dataset}, a data frame with attached meta-data}
}
\value{
object with S3 class \code{stw_dataset}
}
\description{
Create new dataset object
}
