% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stw_meta.R
\name{stw_meta}
\alias{stw_meta}
\title{Create a new meta-data object}
\usage{
stw_meta(name, title, description, source = NULL, n_row = NULL,
  n_col = NULL, dictionary = NULL)
}
\arguments{
\item{name}{\code{character} name of the dataset}

\item{title}{\code{character} title of the dataset}

\item{description}{\code{character} description of the dataset}

\item{source}{\code{character} source of the dataset}

\item{n_row}{\code{integer} number of rows in the dataset}

\item{n_col}{\code{integer} number of columns in the dataset}

\item{dictionary}{\code{stw_dict} object, dictionary of variables in the dataset}
}
\value{
object with S3 class \code{stw_meta}
}
\description{
Create a new meta-data object
}
