% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stw_mutate_dict.R
\name{stw_mutate_dict}
\alias{stw_mutate_dict}
\alias{stw_mutate_dict.default}
\alias{stw_mutate_dict.stw_dict}
\alias{stw_mutate_dict.stw_meta}
\title{Title}
\usage{
stw_mutate_dict(...)

\method{stw_mutate_dict}{default}(...)

\method{stw_mutate_dict}{stw_dict}(dictionary, ...)

\method{stw_mutate_dict}{stw_meta}(meta, ...)
}
\arguments{
\item{...}{Name-value pairs of expressions}

\item{dictionary}{Object with S3 class \code{stw_dict}}

\item{meta}{object with S3 class \code{stw_meta}}
}
\value{
modified copy of \code{dictionary} or \code{meta}
}
\description{
Title
}
\examples{
stw_mutate_dict(diamonds_meta$dictionary, color = "foo")
stw_mutate_dict(diamonds_meta, color = "foo")

}
