% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stw_mutate_dict.R
\name{stw_mutate_dict}
\alias{stw_mutate_dict}
\alias{stw_mutate_dict.default}
\alias{stw_mutate_dict.stw_dict}
\title{Title}
\usage{
stw_mutate_dict(...)

\method{stw_mutate_dict}{default}(...)

\method{stw_mutate_dict}{stw_dict}(dictionary, ...)
}
\arguments{
\item{...}{Name-value pairs of expressions}

\item{dictionary}{Object with S3 class \code{stw_dict}, or \code{stw_meta}. Note that
this need not be a \code{stw_dict} object; it can be an object like \code{stw_meta}
that contains a \code{stw_dict} object.}
}
\value{
modified copy of \code{dictionary}
}
\description{
Title
}
