% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stw_to_roxygen.R
\name{stw_to_roxygen}
\alias{stw_to_roxygen}
\alias{stw_to_roxygen.default}
\alias{stw_to_roxygen.stw_meta}
\alias{stw_use_roxygen}
\alias{stw_write_roxygen}
\title{Output to roxygen format}
\usage{
stw_to_roxygen(x, ...)

\method{stw_to_roxygen}{default}(x, ...)

\method{stw_to_roxygen}{stw_meta}(x, ...)

stw_use_roxygen(meta)

stw_write_roxygen(meta, file)
}
\arguments{
\item{x}{object with S3 class \code{stw_meta} or \code{stw_data}}

\item{file}{\code{character} path of file to write}
}
\value{
\describe{
\item{\code{stw_to_roxygen()}}{\code{character}, roxygen string}
\item{\code{stw_use_roxygen()}}{\code{invisible(x)}, called for side-effects}
\item{\code{stw_write_roxygen()}}{\code{invisible(x)}, called for side-effects}
}
}
\description{
Use these functions to convert your metadata to roxygen format. The function
\code{stw_to_roxygen()} returns a roxygen string, \code{stw_use_roxygen()} copies a
roxygen string to your clipboard, and \code{str_write_yaml()} writes a roxygen
string to a file.
}
\examples{
cat(stw_to_roxygen(diamonds_meta))
stw_write_roxygen(diamonds_meta, tempfile(fileext = ".yml"))
\dontrun{
# not run because it copies text to the clipboard
stw_use_roxygen(diamonds_meta)
}

}
