% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stw_to_roxygen.R
\name{stw_to_roxygen}
\alias{stw_to_roxygen}
\title{Create Roxygen string}
\usage{
stw_to_roxygen(meta)
}
\arguments{
\item{meta}{object with S3 class \code{stw_meta}}
}
\value{
\code{character} Roxygen string
}
\description{
Create Roxygen string
}
