% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stw_to_yaml.R
\name{stw_to_yaml}
\alias{stw_to_yaml}
\title{Create YAML string}
\usage{
stw_to_yaml(meta)
}
\arguments{
\item{meta}{object with S3 class \code{stw_meta}}
}
\value{
object with S3 class \code{glue}, based on character string
}
\description{
Create YAML string
}
\examples{
  str_to_yaml(diamonds_meta)

}
